En proceso de revisión final

+++ src/main/java/com/reelme/reelmespringboot/repository/ResenaRepository.java
@@ -26,7 +26,7 @@ public interface ResenaRepository extends CrudRepository<Resena, Integer> {
 
     int countByFechaBetweenAndNomUsuario(Date dateStart, Date dateEnd, Usuario usuario);
 
-    Resena findById(int id);
+    Resena findById(Long id);
 
     List<Resena> findTop4ByNomUsuarioOrderByFechaDesc(Usuario usuario);
     @Query("SELECT r.idPelicula, COUNT(r) as resenaCount FROM Resena r GROUP BY r.idPelicula ORDER BY resenaCount DESC LIMIT 4")
